payroll-backend/requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django>=5.0
django-cors-headers==4.9.0
django-filter>=24.0
djangorestframework>=3.14
drf-spectacular>=0.27.0
et_xmlfile==2.0.0
fonttools==4.61.1
Markdown==3.10
numpy>=1.26
openpyxl==3.1.5
pandas>=2.2
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.7
ruff==0.14.10
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
psycopg2-binary>=2.9
workalendar==17.0.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
